requirements.txt
psycopg2-binary
pandas
tweepy
requests